Create UK90_height_weight_bmi.csv
This is the data for height, weight and BMI LMS values from the UK90 dataset. I have removed Head Circumference, Sitting Height, Leg Length, Waist and Tanita for performance reasons. The code in UK90algorithm will need only minor alteration though if later on someone wishes to add in the rest of the data.